+++ data/src/config/server/icon.rs
@@ -4,10 +4,14 @@ use serde::Deserialize;
 #[serde(default)]
 pub struct Icon {
     pub enabled: bool,
+    pub override_url: Option<String>,
 }
 
 impl Default for Icon {
     fn default() -> Self {
-        Self { enabled: true }
+        Self {
+            enabled: true,
+            override_url: None,
+        }
     }
 }