+++ src/appearance/theme/svg.rs
@@ -18,6 +18,28 @@ pub fn none(_theme: &Theme, _status: Status) -> Style {
     Style { color: None }
 }
 
+pub fn primary(theme: &Theme, _status: Status) -> Style {
+    Style {
+        color: Some(theme.styles().text.primary.color),
+    }
+}
+
+pub fn unread_indicator(theme: &Theme, _status: Status) -> Style {
+    Style {
+        color: Some(theme.styles().general.unread_indicator),
+    }
+}
+
+pub fn highlight_indicator(theme: &Theme, _status: Status) -> Style {
+    Style {
+        color: theme
+            .styles()
+            .general
+            .highlight_indicator
+            .or(Some(theme.styles().general.unread_indicator)),
+    }
+}
+
 pub fn error(theme: &Theme, _status: Status) -> Style {
     Style {
         color: Some(theme.styles().text.error.color),